web/requirements.txt
Django
psycopg2-binary
django-debug-toolbar
